templates/example_1/requirements.txt
backtrader
pandas
numpy
matplotlib
python-dotenv
